move generateSteps out of microworker
